Moved precondition
